Refactor to put the circuit in a subdirectory.
