+++ gnark/compress/.gitignore
@@ -1,5 +1,3 @@
 /proof-detail-response.json
 /node_modules
-/package.json
-/package-lock.json
 /verify
