+++ gnark/compress/circuit/Sindri.json
