+++ gnark/compress/circuit/circuit.go
