+++ gnark/compress/circuit/go.mod
@@ -1,11 +1,10 @@
-module github.com/sindri-labs/forge-sample-data/gnark/compress
+module github.com/sindri-labs/forge-sample-data/gnark/compress/circuit
 
 go 1.19
 
 require (
 	github.com/consensys/gnark v0.9.0
 	github.com/consensys/gnark-crypto v0.11.2
-	github.com/sindri-labs/forge-sample-data/gnark/compress/circuit v1.1.1
 )
 
 require (
@@ -13,7 +12,6 @@ require (
 	github.com/blang/semver/v4 v4.0.0 // indirect
 	github.com/consensys/bavard v0.1.13 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/fxamacker/cbor/v2 v2.5.0 // indirect
 	github.com/google/pprof v0.0.0-20230817174616-7a8ec2ada47b // indirect
 	github.com/mattn/go-colorable v0.1.13 // indirect
 	github.com/mattn/go-isatty v0.0.19 // indirect
@@ -21,10 +19,7 @@ require (
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	github.com/rs/zerolog v1.30.0 // indirect
 	github.com/stretchr/testify v1.8.4 // indirect
-	github.com/x448/float16 v0.8.4 // indirect
 	golang.org/x/sys v0.11.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 	rsc.io/tmplfunc v0.0.3 // indirect
 )
-
-replace github.com/sindri-labs/forge-sample-data/gnark/compress/circuit => ./circuit
