+++ gnark/compress/circuit/input.json
