+++ gnark/compress/package.json
@@ -1,0 +1,6 @@
+{
+  "dependencies": {
+    "axios": "^1.5.1",
+    "form-data": "^4.0.0"
+  }
+}
